(chores) camel-report-maven-plugin: apply required cast for intermediate variables
Among other things, it should prevent overflows

+++ catalog/camel-report-maven-plugin/src/main/java/org/apache/camel/maven/ValidateMojo.java
@@ -457,7 +457,7 @@ public class ValidateMojo extends AbstractExecMojo {
         int sedaDirectErrors = 0;
         String sedaDirectSummary = "";
         if (directOrSedaPairCheck) {
-            long sedaDirectEndpoints = countEndpointPairs(endpoints, "direct") + countEndpointPairs(endpoints, "seda");
+            long sedaDirectEndpoints = (long) countEndpointPairs(endpoints, "direct") + (long) countEndpointPairs(endpoints, "seda");
             sedaDirectErrors += validateEndpointPairs(endpoints, "direct") + validateEndpointPairs(endpoints, "seda");
             if (sedaDirectErrors == 0) {
                 sedaDirectSummary = String.format("Endpoint pair (seda/direct) validation success: (%s = pairs)", sedaDirectEndpoints);
